DDL Statements -Cakes & Fudges
A case study on a cake shop named Cakes & Fudges. Cakes & Fudges is a mail-order company that serves chocolate delicacies to a large audience of chocolate-crazed customers. Their daring chefs are always innovating, and they need a better way to manage their ever-growing catalog of chocolate goodness. They also need a way to keep track of customers and the orders they make.